Fix performance issue with storing user IP addresses
Originally, I wanted the design of longboard to fully hide user IP
addresses from site administration, to prevent misuse. The way that I
wanted to accomplish this was by hashing user IP addresses and storing
only the hashes, so we could track users by their IP while keeping their
IP addresses completely opaque.

The problem with this is that if you use a cryptographically secure
hash, it has a very negative impact on performance. By nature a
cryptographically secure hash function is slow in order to be resistant
to brute-force attacks. This performance loss is unacceptable and I
can't see a way to keep the current design in a performant way.

Now, the database stores user IP addresses. The hash is still computed
when a new IP address is seen, but if we've already seen an IP the hash
is not re-computed.

Only the hash is displayed to site moderators, but any administrator
which has access to the database can look up the IP for a given user. I
suppose this isn't too bad because anyone who owns the server could, for
example, use longboard behind a reverse proxy like nginx and log user IP
addresses that way. We wouldn't be stopping anything by not storing IP
addresses. The best that we can do is make misuse of IP addresses
inconvenient.

I added a simple benchmark to benches/bench.rs. This issue impacted any
page which took a User parameter, including the homepage, and the fix
cut page load time from several seconds to milliseconds.