Add a systemd .service file
Some of the contrib files were changed to .m4 files, this is so we can
use the values set in the Makefile in them. This way, if you change the
bindir or the resdir when running make or make install, it'll be
reflected in the generated config file and systemd .service file.

Also, some of the default values in config.rs were changed to use env!()
calls to pull the values from the environment, which make supplies.