Write a post parser/formatter
Now, instead of markdown, posts are formatted with our own markdown-like
syntax. The biggest motivation for this is markdown handles linebreaks
differently then you'd want to for posting on an online forum (hitting
the enter key doesn't always insert a text break, which goes against
user expectations).

I still need to write up a guide on what syntax is available. I think
that longboard should ship with a guide included on what every form
option does for new threads/forms, as they're not very self-explanatory.

Two more notes for this commit: I added the ability to create a mock
database, which is useful for situations where you need a database but
you don't need to use it. This is used in the unit tests for the post
parser.

And I added a user option for code highlighting. This links to
Cloudflare's CDN right now, I should probably add a note in the README
about the option of swapping that out for a self-hosted solution, if
admins would prefer that.