Fix incomplete post bug
Because of the way pooled database connections were handled (each new
database operation got a new connection from the pool) database
transactions weren't working. I re-wrote the database code so that each
route now gets only one connection. The Database type was replaced with
a Connection type which is generic (the type aliased PooledConnection
and SinlgeConnection are just aliases for Connection). PooledConnection
implements FromRequest, and this is how request handlers get their
database connection.

This was a big re-write and I'm not sure if all of the work to make it
generic was worth it; longctl _could_ just used pooled connections also.
Still, I think the resulting code is clean and readable. The bug is
fixed which is the important part.

The post parser was changed so that instead of passing a database
connection all the way through the parser chain, post reference
resolving is done at the end. This makes the unit tests simpler to write
because you don't need to pass in a mock database.