Change many things.
First of all, all of the dependencies have been updated to their latest
version. Several of them had breaking changes, which will be discussed
below.

The base-fonts.css file was fixed. Entries must be ended with a
semicolon. I think I didn't notice this before because all of the
webfonts were also installed on my system.

Clap changed a few small things such as the names of functions.
Executables in src/bin/ were updated to reflect these changes.

Some error messages in src/error.rs were updated to be more helpful.

Perhaps the biggest change, diesel database connections now need to be
mutable. All of the Connection methods were changed to reflect this, as
well as most of the constructors in the views in src/view/, which take a
Context. This Context must now be mutable in order to obtain a mutable
reference to the database connection inside of the Context.

Diesel also slightly changed how migrations are implemented, and this is
reflected in src/models/mod.rs.

There are still a couple of things to fix, which I've marked with TODO
or FIXME. These will be addressed in the upcoming commits.