Change how the moderation form works
Before, if you typed a reason into the reason field, the default action
was to delete the post, because that is the first submit button in the
field. This isn't desirable because a moderator might hit enter by
accident or by habit, and accidentally delete a post.

Now, the enter key does nothing in the form. It is required to click the
delete button, or other button such as lock or unpin on threads. Posts
can still be deleted only using the keyboard by tabbing over to the
delete button and hitting enter.