Refactor models
The code for the models has been split up into models/board.rs,
models/thread.rs, models/post.rs, and models/staff.rs.

Additionally, the model functions (implemented as methods on the
Connection type) should now return the proper not found error when
resources aren't found, which means that we don't have to check for that
explicitly in the routing code.