Fix bug related to the maximum file size
Instead of failing silently when a file is too big, now we throw an
error like intended.

Note that the behavior here is slightly different with the way that
multipart form handling is dealing with the max file size. There is no
longer a maximum size for the whole request, only per-file. I wonder if
this could become an avenue for abuse in the future.